requirements.txt
###### Requirements with Version Specifiers #######
certifi==2020.12.5
joblib==1.0.1
scikit-learn==0.24.1
sklearn==0.0
threadpoolctl==2.1.0
wincertstore==0.2
Pillow==8.1.0

###### Requirements without Version Specifiers #######
numpy
keras
scipy
